requirements.txt
streamlit==1.39.0
tensorflow==2.18.1
numpy==1.26.4
Pillow==10.4.0
matplotlib==3.9.2
protobuf==4.25.3
plotly==5.24.1
reportlab==4.2.5
